split also the vem
